+++ src/kvp/consts/values.rs
@@ -1,0 +1,2 @@
+/// The default value for the `stackable.tech/vendor` label.
+pub const STACKABLE_VENDOR_VALUE: &str = "Stackable";