+++ CHANGELOG.md
@@ -15,8 +15,10 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Change Deref target of `KeyPrefix` and `KeyName` from `String` to `str` ([#725]).
+- Add Stackable vendor label `stackable.tech/vendor: Stackable` to recommended labels ([#728]).
 
 [#725]: https://github.com/stackabletech/operator-rs/pull/725
+[#728]: https://github.com/stackabletech/operator-rs/pull/728
 
 ## [0.62.0] - 2024-01-19